i believe i am on the right path -- see https://edstem.org/us/courses/3569/discussion/247463 for my open questions
So i think i have progressed to the point of needing to update parameters

if y ̃ does_not_equal yi
    w ← w + Φ(xi, yi) − Φ(xi, y ̃)

I think I understandΦ(xi, y ̃) -- This would be the vector of features that gave me the estimated NER sequence. Correct?

But what is 'Φ(xi, yi)'? -- Do I get the vector of features that would give me the 'right' NER sequence?
So for example

If i had a feature for POS+NER -- And i know for a given word if it is a Noun -- when i do my Viterbi algorithm, should i be turning on various NERs? So i would try NOUN+O | NOUN+B-ORG | NOUN+I-ORG

Φ(xi, y ̃) would then be the features that gave me the estimated NER Sequence

Φ(xi, yi) would be the features that would have given me the right NER sequence

Aren't some always the same? A Feature like, if word is capitalized, doesn't change in my Viterbi sequence so would be common between Φ(xi, yi) & Φ(xi, y ̃)
Am I on the wrong track?

It could be I have a fundamental misunderstanding still as well -- i have attempted to 'get off the ground' by implementing a single feature for part-of-speech

I track all POS seen in the training data, and then create a feature for each one (~45).

Then i iterate on all sentences again, and if a word has a part of speech, i turn that feature on ({feature_index},1)

I know this will always produce O because when i do viterbi, i end up with the same score for every NER for a word, since features are constant for a word... i don't fully understand when or why a word's viterbi π would differ... which could be where i am confused